database fix

+++ db/olx-database.sql
@@ -4,7 +4,7 @@ GRANT ALL PRIVILEGES ON DATABASE olx TO admin;
 \c olx;
 
 CREATE TABLE USERS (
-                       ID_USER NUMERIC PRIMARY KEY,
+                       ID_USER SERIAL PRIMARY KEY,
                        NAME VARCHAR(255) NOT NULL,
                        SURNAME VARCHAR(255) NOT NULL,
                        PASSWORD VARCHAR(50) NOT NULL,
@@ -14,33 +14,33 @@ CREATE TABLE USERS (
 
 
 CREATE TABLE PROVINCE (
-                          ID_PROVINCE NUMERIC PRIMARY KEY,
+                          ID_PROVINCE SERIAL PRIMARY KEY,
                           NAME VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE CITY (
-                      ID_CITY NUMERIC PRIMARY KEY,
+                      ID_CITY SERIAL PRIMARY KEY,
                       NAME VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE CATEGORY (
-                          ID_CATEGORY NUMERIC PRIMARY KEY,
+                          ID_CATEGORY SERIAL PRIMARY KEY,
                           NAME VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE ADVERTISEMENT (
-                               ID NUMERIC PRIMARY KEY,
+                               ID SERIAL PRIMARY KEY,
                                TITLE VARCHAR(255) NOT NULL,
                                DESCRIPTION VARCHAR(255),
                                PRICE FLOAT,
                                EMAIL VARCHAR(255) NOT NULL,
                                PHONE NUMERIC(20) NOT NULL,
-                               ID_USER NUMERIC NOT NULL,
+                               ID_USER SERIAL NOT NULL,
                                CONSTRAINT FK_ID_USER FOREIGN KEY(ID_USER) REFERENCES USERS(ID_USER),
-                               ID_PROVINCE NUMERIC NOT NULL,
+                               ID_PROVINCE SERIAL NOT NULL,
                                CONSTRAINT FK_ID_PROVINCE FOREIGN KEY(ID_PROVINCE) REFERENCES PROVINCE(ID_PROVINCE),
-                               ID_CITY NUMERIC NOT NULL,
+                               ID_CITY SERIAL NOT NULL,
                                CONSTRAINT FK_ID_CITY FOREIGN KEY(ID_CITY) REFERENCES CITY(ID_CITY),
-                               ID_CATEGORY NUMERIC NOT NULL,
+                               ID_CATEGORY SERIAL NOT NULL,
                                CONSTRAINT FK_ID_CATEGORY FOREIGN KEY(ID_CATEGORY) REFERENCES CATEGORY(ID_CATEGORY)
 );
